// src/types/enums/Estados.ts
export enum Estado {
    PENDIENTE = 'PENDIENTE', // cajero admin
    PREPARACION = 'PREPARACION', // cocinero admin
    TERMINADO = 'TERMINADO', // cajero admin
    RECHAZADO = 'RECHAZADO',  // admin 
    FACTURADO = 'FACTURADO', // admin
    DELIVERY = 'DELIVERY', // empleado
    ENTREGADO = 'ENTREGADO' // admin 
  }
  